Add addComicBook method and collection print functionality to Application class
The addComicBook method allows users to add comic to their collection by providing necessary
information. The ability to print all the comics in a collection allows users to view their
collection contents in their entirety. In regard to the addComicBook method, there is an
opportunity to transition away from an ArrayList to hold ComicBook objects and instead use
a Map.